Struct gooey_core::WindowBuilder

pub struct WindowBuilder<W: Widget> {
    pub initializer: Option<Box<dyn FnOnce(&WidgetStorage) -> StyledWidget<W>>>,
    pub configuration: WindowConfiguration,
}

A builder for a Window.

Fields
initializer: Option<Box<dyn FnOnce(&WidgetStorage) -> StyledWidget<W>>>

The function that creates the root widget for this window.

configuration: WindowConfiguration

The intial configuration of the window.

Implementations

impl<W: Widget> WindowBuilder<W>

pub fn new<F: FnOnce(&WidgetStorage) -> StyledWidget<W> + 'static>(
initializer: F
) -> Self

Creates a new builder with initializer used to create the root widget.

pub fn title(self, title: impl Into<String>) -> Self

Sets the window’s title.

pub fn position(self, location: impl Into<Point<i32, Pixels>>) -> Self

Sets the window’s position (in screen coordinates).

pub fn size(self, size: impl Into<Size<u32, Points>>) -> Self

Sets the window’s size.

pub fn non_resizable(self) -> Self

Prevents the window from being resized.

pub fn maximize(self) -> Self

Maximizes the window upon opening.

pub fn transparent(self) -> Self

Enables transparent window handling, if the platform supports it. Background colors that have transparency will allow other content to show through.

pub fn plain(self) -> Self

Removes decorations from the window (such as the title bar).

pub fn always_on_top(self) -> Self

Sets that the window should stay on top of all other windows.

pub fn open(self, frontend: &dyn AnyFrontend) -> bool

Opens the window. Only possible on platforms that support multiple windows.