Struct gooey_core::style::Name

pub struct Name {
    pub authority: Identifier,
    pub name: Identifier,
}

A globally unique name.

This structure has an authority and a name.

Fields
authority: Identifier

The unique name of the source of this name. For example, this could be the name of the crate it was defined within.

Identifier::private() is used as the authority when Name::private() is called.

name: Identifier

The locally unique name.

This name only needs to be unique within its authority. For example, two authorities can define their own color components without conflicts.

Implementations

impl Name

pub fn private(name: impl Identifiable) -> Result<Name, InvalidIdentifier>

Returns a new Name with _ used as the authority.

This is equivalent to calling Name::new("_", name).

Errors

Returns [InvalidIdentifier] if any invalid Identifier characters are encountered.

pub fn new( authority: impl Identifiable, name: impl Identifiable ) -> Result<Name, InvalidIdentifier>

Returns a new Name using authority and name.

Each name should be unique within the authority namespace.

Errors

Returns [InvalidIdentifier] if any invalid Identifier characters are encountered.