Struct gooey_core::styles::Color

pub struct Color {
    pub red: f32,
    pub green: f32,
    pub blue: f32,
    pub alpha: f32,
}

An Srgba color.

Fields
red: f32

The red channel. Valid values are from 0.0 to 1.0 inclusive.

green: f32

The green channel. Valid values are from 0.0 to 1.0 inclusive.

blue: f32

The blue channel. Valid values are from 0.0 to 1.0 inclusive.

alpha: f32

The alpha channel. Valid values are from 0.0 to 1.0 inclusive, where 1.0 is fully opaque.

Implementations

impl Color

pub const fn new(red: f32, green: f32, blue: f32, alpha: f32) -> Self

Creates a new color with SRGBA components red, green, blue, and alpha ranging from 0.0-1.0.

pub fn new_u8(red: u8, green: u8, blue: u8, alpha: u8) -> Self

Creates a new color with SRGBA components red, green, blue, and alpha ranging from 0-255.

pub fn as_css_string(&self) -> String

Formats the color for CSS.

pub const fn gray(lightness: f32) -> Self

Returns a gray color with lightness in the red, green, and blue channels and an alpha of 1.0.

pub const fn with_alpha(&self, alpha: f32) -> Self

Returns a new color with the same red, green, and blue values, but replacing the current alpha with the parameter.

pub fn lighten(self, factor: f32) -> Self

Returns this color after lightening it by factor. Uses the Laba colorspace for luminance calcluations.

pub fn darken(self, factor: f32) -> Self

Returns this color after darkening it by factor. Uses the Laba colorspace for luminance calcluations.

impl Color

pub const ALICEBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const ANTIQUEWHITE: Self

Equivalent to the CSS color keywords of the same name.

pub const AQUA: Self

Equivalent to the CSS color keywords of the same name.

pub const AQUAMARINE: Self

Equivalent to the CSS color keywords of the same name.

pub const AZURE: Self

Equivalent to the CSS color keywords of the same name.

pub const BEIGE: Self

Equivalent to the CSS color keywords of the same name.

pub const BISQUE: Self

Equivalent to the CSS color keywords of the same name.

pub const BLACK: Self

Equivalent to the CSS color keywords of the same name.

pub const BLANCHEDALMOND: Self

Equivalent to the CSS color keywords of the same name.

pub const BLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const BLUEVIOLET: Self

Equivalent to the CSS color keywords of the same name.

pub const BROWN: Self

Equivalent to the CSS color keywords of the same name.

pub const BURLYWOOD: Self

Equivalent to the CSS color keywords of the same name.

pub const CADETBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const CHARTREUSE: Self

Equivalent to the CSS color keywords of the same name.

pub const CHOCOLATE: Self

Equivalent to the CSS color keywords of the same name.

pub const CLEAR_BLACK: Self

Equivalent to the CSS color keywords of the same name.

pub const CLEAR_WHITE: Self

Equivalent to the CSS color keywords of the same name.

pub const CORAL: Self

Equivalent to the CSS color keywords of the same name.

pub const CORNFLOWERBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const CORNSILK: Self

Equivalent to the CSS color keywords of the same name.

pub const CRIMSON: Self

Equivalent to the CSS color keywords of the same name.

pub const CYAN: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKCYAN: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKGOLDENROD: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKGRAY: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKGREEN: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKGREY: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKKHAKI: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKMAGENTA: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKOLIVEGREEN: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKORANGE: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKORCHID: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKRED: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKSALMON: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKSEAGREEN: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKSLATEBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKSLATEGRAY: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKSLATEGREY: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKTURQUOISE: Self

Equivalent to the CSS color keywords of the same name.

pub const DARKVIOLET: Self

Equivalent to the CSS color keywords of the same name.

pub const DEEPPINK: Self

Equivalent to the CSS color keywords of the same name.

pub const DEEPSKYBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const DIMGRAY: Self

Equivalent to the CSS color keywords of the same name.

pub const DIMGREY: Self

Equivalent to the CSS color keywords of the same name.

pub const DODGERBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const FIREBRICK: Self

Equivalent to the CSS color keywords of the same name.

pub const FLORALWHITE: Self

Equivalent to the CSS color keywords of the same name.

pub const FORESTGREEN: Self

Equivalent to the CSS color keywords of the same name.

pub const FUCHSIA: Self

Equivalent to the CSS color keywords of the same name.

pub const GAINSBORO: Self

Equivalent to the CSS color keywords of the same name.

pub const GHOSTWHITE: Self

Equivalent to the CSS color keywords of the same name.

pub const GOLD: Self

Equivalent to the CSS color keywords of the same name.

pub const GOLDENROD: Self

Equivalent to the CSS color keywords of the same name.

pub const GRAY: Self

Equivalent to the CSS color keywords of the same name.

pub const GREEN: Self

Equivalent to the CSS color keywords of the same name.

pub const GREENYELLOW: Self

Equivalent to the CSS color keywords of the same name.

pub const GREY: Self

Equivalent to the CSS color keywords of the same name.

pub const HONEYDEW: Self

Equivalent to the CSS color keywords of the same name.

pub const HOTPINK: Self

Equivalent to the CSS color keywords of the same name.

pub const INDIANRED: Self

Equivalent to the CSS color keywords of the same name.

pub const INDIGO: Self

Equivalent to the CSS color keywords of the same name.

pub const IVORY: Self

Equivalent to the CSS color keywords of the same name.

pub const KHAKI: Self

Equivalent to the CSS color keywords of the same name.

pub const LAVENDER: Self

Equivalent to the CSS color keywords of the same name.

pub const LAVENDERBLUSH: Self

Equivalent to the CSS color keywords of the same name.

pub const LAWNGREEN: Self

Equivalent to the CSS color keywords of the same name.

pub const LEMONCHIFFON: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTCORAL: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTCYAN: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTGOLDENRODYELLOW: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTGRAY: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTGREEN: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTGREY: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTPINK: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTSALMON: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTSEAGREEN: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTSKYBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTSLATEGRAY: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTSLATEGREY: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTSTEELBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const LIGHTYELLOW: Self

Equivalent to the CSS color keywords of the same name.

pub const LIME: Self

Equivalent to the CSS color keywords of the same name.

pub const LIMEGREEN: Self

Equivalent to the CSS color keywords of the same name.

pub const LINEN: Self

Equivalent to the CSS color keywords of the same name.

pub const MAGENTA: Self

Equivalent to the CSS color keywords of the same name.

pub const MAROON: Self

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMAQUAMARINE: Self

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMORCHID: Self

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMPURPLE: Self

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMSEAGREEN: Self

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMSLATEBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMSPRINGGREEN: Self

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMTURQUOISE: Self

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMVIOLETRED: Self

Equivalent to the CSS color keywords of the same name.

pub const MIDNIGHTBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const MINTCREAM: Self

Equivalent to the CSS color keywords of the same name.

pub const MISTYROSE: Self

Equivalent to the CSS color keywords of the same name.

pub const MOCCASIN: Self

Equivalent to the CSS color keywords of the same name.

pub const NAVAJOWHITE: Self

Equivalent to the CSS color keywords of the same name.

pub const NAVY: Self

Equivalent to the CSS color keywords of the same name.

pub const OLDLACE: Self

Equivalent to the CSS color keywords of the same name.

pub const OLIVE: Self

Equivalent to the CSS color keywords of the same name.

pub const OLIVEDRAB: Self

Equivalent to the CSS color keywords of the same name.

pub const ORANGE: Self

Equivalent to the CSS color keywords of the same name.

pub const ORANGERED: Self

Equivalent to the CSS color keywords of the same name.

pub const ORCHID: Self

Equivalent to the CSS color keywords of the same name.

pub const PALEGOLDENROD: Self

Equivalent to the CSS color keywords of the same name.

pub const PALEGREEN: Self

Equivalent to the CSS color keywords of the same name.

pub const PALETURQUOISE: Self

Equivalent to the CSS color keywords of the same name.

pub const PALEVIOLETRED: Self

Equivalent to the CSS color keywords of the same name.

pub const PAPAYAWHIP: Self

Equivalent to the CSS color keywords of the same name.

pub const PEACHPUFF: Self

Equivalent to the CSS color keywords of the same name.

pub const PERU: Self

Equivalent to the CSS color keywords of the same name.

pub const PINK: Self

Equivalent to the CSS color keywords of the same name.

pub const PLUM: Self

Equivalent to the CSS color keywords of the same name.

pub const POWDERBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const PURPLE: Self

Equivalent to the CSS color keywords of the same name.

pub const REBECCAPURPLE: Self

Equivalent to the CSS color keywords of the same name.

pub const RED: Self

Equivalent to the CSS color keywords of the same name.

pub const ROSYBROWN: Self

Equivalent to the CSS color keywords of the same name.

pub const ROYALBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const SADDLEBROWN: Self

Equivalent to the CSS color keywords of the same name.

pub const SALMON: Self

Equivalent to the CSS color keywords of the same name.

pub const SANDYBROWN: Self

Equivalent to the CSS color keywords of the same name.

pub const SEAGREEN: Self

Equivalent to the CSS color keywords of the same name.

pub const SEASHELL: Self

Equivalent to the CSS color keywords of the same name.

pub const SIENNA: Self

Equivalent to the CSS color keywords of the same name.

pub const SILVER: Self

Equivalent to the CSS color keywords of the same name.

pub const SKYBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const SLATEBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const SLATEGRAY: Self

Equivalent to the CSS color keywords of the same name.

pub const SLATEGREY: Self

Equivalent to the CSS color keywords of the same name.

pub const SNOW: Self

Equivalent to the CSS color keywords of the same name.

pub const SPRINGGREEN: Self

Equivalent to the CSS color keywords of the same name.

pub const STEELBLUE: Self

Equivalent to the CSS color keywords of the same name.

pub const TAN: Self

Equivalent to the CSS color keywords of the same name.

pub const TEAL: Self

Equivalent to the CSS color keywords of the same name.

pub const THISTLE: Self

Equivalent to the CSS color keywords of the same name.

pub const TOMATO: Self

Equivalent to the CSS color keywords of the same name.

pub const TURQUOISE: Self

Equivalent to the CSS color keywords of the same name.

pub const VIOLET: Self

Equivalent to the CSS color keywords of the same name.

pub const WHEAT: Self

Equivalent to the CSS color keywords of the same name.

pub const WHITE: Self

Equivalent to the CSS color keywords of the same name.

pub const WHITESMOKE: Self

Equivalent to the CSS color keywords of the same name.

pub const YELLOW: Self

Equivalent to the CSS color keywords of the same name.

pub const YELLOWGREEN: Self

Equivalent to the CSS color keywords of the same name.