Trait gooey_core::AnyFrontend

pub trait AnyFrontend: AnySendSync {
    #[must_use]
    fn cloned(&self) -> Box<dyn AnyFrontend>;

    #[must_use]
    fn storage(&self) -> &WidgetStorage;

    fn window(&self) -> Option<&dyn Window>;

    fn theme(&self) -> SystemTheme;

    fn set_widget_has_messages(&self, widget: WidgetId);

    #[must_use]
    fn enter_managed_code(&self) -> ManagedCodeGuard;

    fn load_image(
        &self, 
        asset: &Image, 
        completed: Callback<Image>, 
        error: Callback<String>
    );

    fn asset_url(&self, asset: &Asset) -> Option<Url>;

    fn schedule_timer(
        &self, 
        callback: Callback, 
        duration: Duration, 
        repeating: bool
    ) -> Timer;

    #[must_use]
    fn localize<'a>(
        &self, 
        key: &str, 
        parameters: Option<LocalizationParameters<'a>>
    ) -> String;

    fn app(&self) -> &AppContext { ... }
}

An interface for Frontend that doesn’t requier knowledge of associated types.

Required methods

#[must_use]
fn cloned(&self) -> Box<dyn AnyFrontend>

Clones the frontend, returning the clone in a box.

#[must_use]
fn storage(&self) -> &WidgetStorage

Returns the widget storage.

fn window(&self) -> Option<&dyn Window>

Returns the window for this frontend instance.

fn theme(&self) -> SystemTheme

Returns the current system theme.

fn set_widget_has_messages(&self, widget: WidgetId)

Notifies the frontend that a widget has messages. Frontends should ensure that process_widget_messages is called at some point after this method is called.

#[must_use]
fn enter_managed_code(&self) -> ManagedCodeGuard

Marks that managed code is being executed. Can be nested. Automatically exited when the returned guard is dropped.

fn load_image(
&self,
asset: &Image,
completed: Callback<Image>,
error: Callback<String>
)

Loads an image asynchronously, executing completed when loaded.

fn asset_url(&self, asset: &Asset) -> Option<Url>

Returns the full Url for the asset, if available.

fn schedule_timer(
&self,
callback: Callback,
duration: Duration,
repeating: bool
) -> Timer

Schedules a timer that invokes callback after duration, and repeats if repeating is true.

#[must_use]
fn localize<'a>(
&self,
key: &str,
parameters: Option<LocalizationParameters<'a>>
) -> String

Localizes key with parameters.

Provided methods

fn app(&self) -> &AppContext

Returns the current application context.