Trait gooey_core::Frontend

pub trait Frontend: Clone + Debug + Send + Sync + 'static {
    type AnyTransmogrifier: AnyTransmogrifier<Self>;
    type Context;

    fn gooey(&self) -> &Gooey<Self>;

    fn theme(&self) -> SystemTheme;

    fn set_widget_has_messages(&self, widget: WidgetId);

    fn load_image(
        &self, 
        asset: &Image, 
        completed: Callback<Image>, 
        error: Callback<String>
    );

    fn asset_configuration(&self) -> &Configuration;

    fn schedule_timer(
        &self, 
        callback: Callback, 
        duration: Duration, 
        repeating: bool
    ) -> Timer;

    fn window(&self) -> Option<&dyn Window>;

    fn open(&self, window: Box<dyn AnyWindowBuilder>) -> bool;

    fn ui_state_for(&self, widget_id: &WidgetId) -> State { ... }

    fn asset_url(&self, asset: &Asset) -> Option<Url> { ... }

    fn exit_managed_code(&self) { ... }

    fn widget_initialized(&self, widget: &WidgetId, style: &Style) { ... }

    fn localize<'a>(
        &self, 
        key: &str, 
        parameters: impl Into<Option<LocalizationParameters<'a>>>
    ) -> String { ... }

}

A frontend is an implementation of widgets and layouts.

Associated Types

type AnyTransmogrifier: AnyTransmogrifier<Self>

The generic-free type of the frontend-specific transmogrifier trait.

type Context

The context type provided to aide in transmogrifying.

Required methods

fn gooey(&self) -> &Gooey<Self>

Returns the underlying Gooey instance.

fn theme(&self) -> SystemTheme

Returns the current system theme.

fn set_widget_has_messages(&self, widget: WidgetId)

Notifies the frontend that a widget has messages. Frontends should ensure that process_widget_messages is called at some point after this method is called.

fn load_image(
&self,
asset: &Image,
completed: Callback<Image>,
error: Callback<String>
)

Loads an image asynchronously, executing completed when loaded.

fn asset_configuration(&self) -> &Configuration

Returns the asset configuration.

fn schedule_timer(
&self,
callback: Callback,
duration: Duration,
repeating: bool
) -> Timer

Schedules a timer that invokes callback after duration, and repeats if repeating is true.

fn window(&self) -> Option<&dyn Window>

Returns the window for this interface, if present.

fn open(&self, window: Box<dyn AnyWindowBuilder>) -> bool

Opens a window. Returns false if unable to open the window.

Provided methods

fn ui_state_for(&self, widget_id: &WidgetId) -> State

Returns the State for the widget. Not all frontends track UI state, and this function is primarily used when building frontends.

fn asset_url(&self, asset: &Asset) -> Option<Url>

Returns the full Url for the asset, if available.

fn exit_managed_code(&self)

Executed when Gooey exits a managed code block.

fn widget_initialized(&self, widget: &WidgetId, style: &Style)

A widget is being initialized.

fn localize<'a>(
&self,
key: &str,
parameters: impl Into<Option<LocalizationParameters<'a>>>
) -> String

Localizes key with parameters.