Trait gooey_core::RelatedStorage

pub trait RelatedStorage<K: Key>: Debug + Send + Sync + 'static {
    fn widget(&self) -> WeakWidgetRegistration;

    fn remove(&self, key: &K);

    fn register(&self, key: K, widget: &WidgetRegistration);
}

Related storage enables a widget to communicate in a limited way about widgets being inserted or removed.

Required methods

fn widget(&self) -> WeakWidgetRegistration

Returns the registration of the widget that this is from.

fn remove(&self, key: &K)

Removes the widget with key from this storage.

fn register(&self, key: K, widget: &WidgetRegistration)

Registers widget with key.

Loading content...