Trait gooey_core::AnyFrontend

pub trait AnyFrontend: AnySendSync {
#[must_use]    fn cloned(&self) -> Box<dyn AnyFrontend>;

#[must_use]    fn storage(&self) -> &WidgetStorage;

    fn set_widget_has_messages(&self, widget: WidgetId);

#[must_use]    fn enter_managed_code(&self) -> ManagedCodeGuard;
}

An interface for Frontend that doesn’t requier knowledge of associated types.

Required methods

#[must_use]fn cloned(&self) -> Box<dyn AnyFrontend>

Clones the frontend, returning the clone in a box.

#[must_use]fn storage(&self) -> &WidgetStorage

Returns the widget storage.

fn set_widget_has_messages(&self, widget: WidgetId)

Notifies the frontend that a widget has messages. Frontends should ensure that process_widget_messages is called at some point after this method is called.

#[must_use]fn enter_managed_code(&self) -> ManagedCodeGuard

Marks that managed code is being executed. Can be nested. Automatically exited when the returned guard is dropped.

Loading content...