Trait gooey_core::Transmogrifier

pub trait Transmogrifier<F: Frontend> {
    type Widget: Widget;
    type Context;
    fn content_size(
        &self, 
        widget: &Self::Widget, 
        constraints: Size2D<Option<f32>, Points>, 
        context: &Self::Context
    ) -> Size2D<f32, Points>;
}

Transforms a Widget into whatever is needed for Frontend F.

Associated Types

type Widget: Widget

The type of the widget being transmogrified.

type Context

The frontend-specific context type provided to aide in transmogrifying.

Loading content...
Required methods

fn content_size(
&self,
widget: &Self::Widget,
constraints: Size2D<Option<f32>, Points>,
context: &Self::Context
) -> Size2D<f32, Points>

Calculate the content-size needed for this widget, trying to stay within constraints.

Loading content...