Trait gooey_core::Transmogrifier

pub trait Transmogrifier<F: Frontend>: Debug + Sized {
    type Widget: Widget;
    type State: Default + Debug + Any + Send + Sync;
    fn initialize(
        &self, 
        widget: &mut Self::Widget, 
        reference: &WidgetRef<Self::Widget>, 
        frontend: &F
    ) -> Self::State { ... }

    fn receive_command(
        &self, 
        command: <Self::Widget as Widget>::Command, 
        context: &mut TransmogrifierContext<'_, Self, F>
    ) { ... }

    fn process_messages(
        &self, 
        transmogrifier_context: TransmogrifierContext<'_, Self, F>
    ) { ... }

    fn default_state_for(
        &self, 
        widget: &mut Self::Widget, 
        reference: &WidgetRef<Self::Widget>, 
        frontend: &F
    ) -> TransmogrifierState { ... }

    fn widget_type_id(&self) -> TypeId { ... }
}

Transforms a Widget into whatever is needed for Frontend F.

Associated Types

type Widget: Widget

The type of the widget being transmogrified.

type State: Default + Debug + Any + Send + Sync

The type the storage this transmogrifier uses for state.

Loading content...
Provided methods

fn initialize(
&self,
widget: &mut Self::Widget,
reference: &WidgetRef<Self::Widget>,
frontend: &F
) -> Self::State

Called after a transmogrifier is initialized.

fn receive_command(
&self,
command: <Self::Widget as Widget>::Command,
context: &mut TransmogrifierContext<'_, Self, F>
)

Called when a command is received from the widget.

fn process_messages(
&self,
transmogrifier_context: TransmogrifierContext<'_, Self, F>
)

Processes commands and events for this widget and transmogrifier.

fn default_state_for(
&self,
widget: &mut Self::Widget,
reference: &WidgetRef<Self::Widget>,
frontend: &F
) -> TransmogrifierState

Returns an initialized state using Self::State::default().

fn widget_type_id(&self) -> TypeId

Returns the TypeId of Self::Widget.

Loading content...