Trait gooey_core::Widget

pub trait Widget: Debug + Send + Sync + Sized + 'static {
    type Command: Debug + Send + Sync;
    type Event: Debug + Send + Sync;

    const CLASS: &'static str;
    fn receive_event(&mut self, event: Self::Event, context: &Context<Self>) { ... }

    #[must_use]
    fn text_color(style: &Style) -> Option<&ColorPair> { ... }

    #[must_use]
    fn background_color(style: &Style) -> Option<&ColorPair> { ... }
}

A graphical user interface element.

Associated Types

type Command: Debug + Send + Sync

Widgets may need to communicate with transmogrifier implementations. This type is the type that can be sent to a transmogrifier.

type Event: Debug + Send + Sync

The type of the event that any Transmogrifier for this widget to use.

Loading content...
Associated Constants

const CLASS: &'static str

The unique class name for this widget. Must not conflict with any other widgets in use. Widget authors should prefix their widget names to ensure no conflicts. For example, the gooey-widgets crate prefixes all of the CLASS constants with gooey-.

Loading content...
Provided methods

fn receive_event(&mut self, event: Self::Event, context: &Context<Self>)

Called when an event from the transmogrifier was received.

#[must_use]
fn text_color(style: &Style) -> Option<&ColorPair>

Returns the effective text color for the given style.

#[must_use]
fn background_color(style: &Style) -> Option<&ColorPair>

Returns the effective background color for the given style.

Loading content...